Using Bootstrap's Grid to tile buttons
For that I nested the three buttons in a div giving it a class row and in each button giving a div with the class col-xs-4.

The 'col' represents the column, the 'xs' to define a grid for devices that have a screen smaller than 768px and the 4 represents the number of columns of width that the buttons must have.