Creating a header for the Cat Photo App
For the header we want the title and the first image to be on the same line.

We use Bootstrap's grid system to help with this problem.

We nest the h2 element and the first image within a single <div class=row> element.

The h2 element being nested within a <div class=col-xs-8> .

The image being nested in a <div class=col-xs-4>.