Adding Information Icon and Trash Icon with Font Awesome
We use the info icon for the info button and the trash can icon for the delete button.

We added the i element to the info button that we gave the fas fa-info-circle classes to it.

We made another i element for the delete button and gave it the fas fa-trash classes.