Putting the input and submit button on the same line
In order for the input and the button to be on the same line, we nest these two elements the input and the button in a div giving it the row class.

In the input element, we add a div with the col-xs-7 class.

In the button element we add a div with the col-xs-5 class.