Labeling Bootstrap wells
To make it clear how the wells are organized, we will name the two wells with their respective ID's.

Above the div with the class well and inside the col-xs-6 div, add the tag h4 and it gives the text that will be #left-well referencing the first well which is the left well.

Above the div with class well and inside the div and inside the col-xs-6 div, add the h4 tag and it gives the text that will be #right-well referencing the right well.