Fit the h2, img and button elements inside your divs
Adjust the title div block by placing the h2 tag inside the div with the col-xs-8 class.

Adjust the div block that has the col-xs-4 class by placing the first img element inside that div.

Adjust the block with the three divs that have the col-xs-4 class and inside them keep the Like, Info and Delete buttons.