Adjust the div blocks from within the container-fluid class
Adjust the first two div blocks that have the col-xs-8 and col-xs-4 classes and place the h2 and a elements inside them.

Also adjust the three div blocks that have the col-xs-4 classes and place inside them the button element with the texts of Like, Info and Delete.

Adjust the div blocks that have the col-xs-6 and col-xs-4 classes from within the form, make the label tags stay within these respective classes.