Adjust div blocks from within the div with the container-fluid class
Adjust the first two divs inside the div class container-fluid that have classes col-xs-8 and col-xs-4 make sure the h2 and a elements are inside these two divs.

Adjust the other three divs that have the col-xs-4 classes to make sure that the button elements with the text Like, Info and Delete are inside those three divs.

Adjust the div blocks that are inside the form and make sure the label tags are inside each of these divs.