Add Flappy Classes to HTML Elements with jQuery
Select the div elements that have the class well and add to them the function addClass( ) and add to them an animated and fluttering class making these elements vibrate animatedly when loading the page.

Select the elements that have the text-primary class and add the animated shake class to those elements so that they also have the same vibration when loading the HTML document.