Select HTML elements by id attribute with jQuery
Inside the document function, select the id with the # of the button element that has the id target3 and target6.

Then, use the addClass( ) function and add the animated fadeOut class, giving this class to these two button elements, it will make these two buttons disappear from the screen when loading the HTML document.