Target same element with multiple jQuery selectors
Select the same element but with three different types: tag, class and id.

Select the button element and give it the animated class, all button elements will have this class.

Select the .btn class and give it the shake class, this makes all buttons with this class shake when loading the HTML document.

Select the id #target1 attribute and give it the btn-primary class, this makes the button element that has that id have the blue backround.