Select the parent of an element with the jQuery parent function
Every HTML element has a parent element from which it inherits its properties.

jQuery has a function called .parent( ) that allows you to access the parent element of that passed element.

Chain the parent() function with the css() function to select the parent of the #target1 element and give the parent element a red background.