Select child elements with jQuery children function
Child elements are those elements that are placed one level below the other.

jquery has a function called .children( ) that allows accessing the child element or elements of that passed element.

Chain the children( ) and css( ) functions to select the child elements of the right-well div and to set the orange color to their text.