Select specific child elements with jQuery's target:nth-child(n)
In jQuery it is possible to use CSS selectors, selecting elements by their classes, id or by their type.

The target:nth-child(n) selector allows you to select through all child elements that are in the nth child of the parent element.

Select the second child elements of the wells to jump when loading the HTML document.