Adjust the HTML structure of the page
The HTML document does not have the complete structure of an HTML.

Make the necessary adjustments so that the page's HTML document has the complete HTML structure.

Remove the button elements from the div blocks, leaving only one button element and create an input and label tag for the page to have a field for the user to choose a background color for the body.

At the end of the body, add the script element referencing a script file that we will be using.