Use jQuery functions and selectors to manipulate the page's HTML elements
To use the jQuery library, within the head tag of the HTML we will place the jQuery library link so that we can use it in the Javascript file.

We will use the jQuery framework to handle and give events to HTML elements.

How to give the body a background, align the text of the h3 element to the center, get the input color value to change the background color when the user clicks on the button.