Use the SASS extension on the page
To do this, install SASS in VSCode extensions and create a styles.scss file.

Inside that file put the styles for the header, for the blog-post class and the h2 element.

SASS has a feature that it is possible to store data using variables, variables are defined with a $ at the beginning and the name of the variable.

Create a variable called text-color and store the red color to it and use this variable to give the red color to the blog-post class and the h2 element.