Nest the h1 and p child elements inside the blog-post parent element with Sass
Sass allows the style rules of child elements to be nested within a parent element in CSS making it easier to organize the style sheet.

Unlike what we are used to seeing an element being directed to a different line.

Use the technique to nest the child elements of the parent .blog-post element with the h1 element coming before the p element.