Use the mixin for CSS reuse
In sass, it is possible to color rules of style of several elements for a mixin function and in this function to put the properties that the element will have and later to include this mixin function inside an element.

First you need to put in front the @mixin keyword and a personalized name, the parameters start with the $ symbol and they are optional.

For a mixin function to be called, you must pass the @include keyword with the custom name of the mixin function and if the function has parameters, you must pass the arguments to it.

Write a mixin called border-radius with a  parameter, the mixin function will have the prefixes: -webkit-border-radius, -moz-border-radius, -ms-border-radius and border-radius with the values of argument that is passed to the function.

Include the mixin function in the div with the id awesome attribute, passing it 15px as an argument.