Add logic to your styles using @if and @else
Just as in Javascript we have if and else statements, we also have in Sass where we can add logic to our styles.

In Sass, we use them with the @ symbol in front of them.

We can also put more than one statement in Sass with @else if and @else just like in Javascript.

Create a mixin called border-stroke with a  parameter, the mixin function checking the following conditions:

Using the @if check if the parameter value  is light set the border 1px solid black.

Using the @else if check if the parameter value  is medium set the border 3px solid black.

Using another @else if check if the value of the  parameter is heavy set the border to 6px solid black.

Lastly, using @else if none of the above conditions satisfy, set the border to none.