Create a Sass loop using @for
In Sass, you can add styles to elements in a loop just like in Javascript.

The Sass loop syntax starts with @for followed by a variable that will represent the index, the difference in this loop is that it has the words from and through, from represents where this variable will start the loop and through represents how far it goes including final.

Use the @for loop that receives the variable  that goes from 1 to 6, this loop should create 6 classes called text-1 to text-6 where each one has a font size defined as 15px multiplied by the index.