Apply a style while the condition is true using the Sass @while loop
The @while loop in Sass is similar to the while loop in Javascript, it creates CSS style rules as long as the condition is true.

Start a variable x with the value 1 and use the @while directive to create a series of classes with different sizes of fonts, the size of these fonts must be defined 15px multiplied by the current number of the index.

Increment the  variable by 1 to avoid the infinite loop.