Use partials to import separate files with code in SASS
Partials are separate files that contain CSS rules within them, it is possible to use them in SASS files.

Partial names start with _ followed by the file name and .scss extension.

If you want to use a partial code for a SASS file, just import that file with the word @import followed by whatever you want to import into the SASS file.

For example, to import all mixin functions that are inside a partial named _mixins.scss, use this in the SASS file: @import mixins.

Use the @import directive to import a partial called _variables.scss into the main.css file without the leading underscore and extension, just the variables.

Once this is done, the main.css will have all the variables and other code available to use.