Use extend SASS to extend a set of styles from one element into another element.
Sass has a feature that allows you to take a set of CSS rules from one element and reuse them in another element, this feature is called @extend.

When using this feature, this other element will have the same properties as an element and can add new styles.

Make the .info-important class extend from the .info class and have a background color set to margin.