Render a div element with its elements inside a JSX variable
One important thing in JSX is that for the browser to recognize the code we need to render the HTML elements.

To do this, define a constant variable and render the div that will contain the elements h1, p, ul, and three li, by doing this the JSX can be transpiled.