Add comments in JSX
JSX is a syntax compiled into valid Javascript. As in other languages, comments in a code are used to make it easier to understand what that code is doing there.

In JSX you can also use comments, to make comments in JSX, you use the syntax {/* */} to wrap the comment text.

Add a comment somewhere inside the div element, without modifying or altering the existing h1 and p elements.