Render HTML elements to the DOM via ReactDOM.render()
React has a rendering API that allows you to render JSX in the HTML DOM, this API is called ReactDOM.

ReactDOM offers a method called render( ), it renders React elements into the DOM, the syntax is like this ReactDOM.render( ).

ReactDOM.render() takes two parameters, the first argument is the React element or component you want to render and the second argument is the DOM node representing which element you want to render the component.

Use the ReactDOM.render() method and pass as the first argument the JSX component for this component to be rendered on the page and as the second parameter use the document.getElementById() and select the DOM node challenge-node to be rendered as well.