Create a functional stateless component
Components are often used for creating complex pages, it is a way for you to organize your page through separate files with the necessary elements to have a good experience in the User Interface and also the maintenance of the page.

Navigation menus, buttons, headers and other content on a page can be created using components.

There are two ways to create React components, the first way is to use a Javascript function, this function will have to return a JSX or null, the name of the function needs to start with a capital letter, the moment it is created, the component will be stateless.

The second way to create components is using classes.

Create a component with a function called MyComponent, this function will return a div element with any text.