Create a React component with classes
Another way to create components is using the same ES6 classes syntax, using the word class followed by the name of that class, then we use the word extends saying from whom we will inherit characteristics and behaviors of React.Component.

Inside the created class we will use the constructor function passing props and calling a component constructor with the word super also passing props to this component and getting the values of the properties of the object instantiated by this class, doing this the component is created correctly.

Finish creating the component with the render method that returns a div element that contains an h1 element with the text Hello React.