Render nested components in React
One of the powerful features that React has is component composition.

In the Apps that we are going to create, we have to think about the user interface. One way to make this easier is to build smaller blocks, with each of these blocks having an important responsibility on our page.

It could be creating a component for your page button, also a component for the navigation menu or for the page footer. In React, it allows you to unite these components into a parent component.

Take the functional component called TypesOfFruit and place it inside the Fruits component.

Now take the Fruits component and nest it inside the TypesOfFood component.