Render a Class Component to the DOM
No React code you write will be rendered in the DOM without using the ReactDOM API call, it's just that React components are passed in the API with a different syntax than JSX elements.

For JSX elements you just need to put the name of the element you want to render, for React components you will need to use the same syntax as if you were rendering a nested component, you will use this syntax for class components and functional components:

ReactDOM.render(<ComponentToRender />, targetNode>).

Nest the Fruits and Vegetables components inside the TypesOfFood component and then use the ReactDOM API with the render method to render the TypesOfFood component in the DOM and we want to render this component inside the div with the id challenge-node attribute.