Create a component from scratch
Define a class called MyComponent that extends from React.Component, with that component's constructor function with the word super inside it, and passing props to both.

Render this component to the DOM using ReactDOM.render( ) and pass a div element with the id challenge-node attribute, referencing where we will render this component.