Pass props to CurrentDate child component
In React, you are allowed to pass props or properties to stateless functional child components, you just need to pass the props as an argument of the functional component.

If you use this component somewhere in the parent component, you use the component name and then a property name as an HTML attribute and pass the value of that property.

With this we will be able to access the value of this property within a tag.

When rendering the CurrentDate in the Calendar component, pass a property called date and give this property a value that will be a Javascript Date object.

Then, access this property in the CurrentDate component showing its value in the p tags, to get the values of the props you need to put the property value between braces, for example date= {Date()}.