Pass an array as props
To pass an array as props to a component, you must use as in Javascript and put curly braces around the array:

prop: {[ item1, item2, item3 .....]}.

Accessing this property is possible using array methods, if you want to join the items to form a string use the join() method.

Create two instances of the List component, both will have a property called tasks and their value will be an array of tasks.

In the first instance of the List class component, give as value an array of pending tasks ['walk dog', 'workout'] under the h2 element with the text Today, representing today's tasks.

In the second instance, give as value an array with three tasks below the h2 element with the text tomorrow, representing tomorrow's tasks.

Once that's done, access the array of tasks in the List component, showing their values in the p element, use the join() method to display the comma-separated lists of tasks.