Use standard props on components
React has an option to define default props or properties, you can assign default props to a component as a property of that component and React will assign the default prop.

React assigns default props if the props are undefined, but if you pass a null value for this prop, it will be null.

Set the ShoppingCart component's props as standard props and specify items as a property passing the value 0 to it.