Replace default props by setting their values
Defining default props is a very useful feature, the way you can override default props is to define their values for a component.

The ShoppingCart component that renders a child component called Items and this component Items has a default quantity prop, it is set to 0.

Replace this default quantity prop by passing the value 10, as the quantity value is an integer, it will not be enclosed in quotes, but must be enclosed in curly braces, like this: {10}.