Use PropTypes to check the type of props value
React provides a feature that allows you to check if component props receive correct type values, for example if you are making an application and then your application makes an API call to retrieve important user data and you expect that data are in an array, they will be passed to your component as a prop.

For this, you will need to use propTypes in your component and require that the data be of type array.

Define propTypes for the Items component and require the quantity prop value to be of type number.