Access a component's props using the this keyword
Whenever you refer to a class component and access that component's prop, use the keyword this followed by props.

For example, if an ES6 class component has a prop called data, you write {this.props.data} in JSX.

Render the Welcome component instance within the parent App component. Give the Welcome component a prop called name and give this prop a string.

Then, inside the Welcome child component, access the prop name and show the value of the prop name in the strong tag.