Remove the use of props in parameters of stateless React components
props is a shorthand way of saying properties, they are used to pass data between components.

A stateless component in React is a class that extends from React.Component, but it doesn't use its internal state.

A common pattern in React is to avoid state ownership as much as possible and create stateless functional components.

Define the Camper component and give it the default props called { name: 'CamperBot' } and inside the Camper, render whatever code you want, but make sure you have a p element accessing that prop name and showing its value.

Define propTypes for the Camper and check if the name property value is of type string.