Create a state property for the component
State is one of the most interesting topics in React, because the state of data and information in web applications changes all the time and new information and data arrives, you want to have your application always responding to this information in an updated way in the UI.

React offers a solution for managing the state of web applications.

You create state in a React component by declaring a state property on the component class within its constructor function.

The state property is defined as a Javascript object and with the this keyword in front of it, like this:

this.state = {

}

The StatefulComponent class component is trying to render the firstName property of its state. However, there is no defined state.

Create the state property on the class component called StatefulComponent, it should be initialized with a firstName property set to a string.