Display a property value from within state in the UI
Once you've set the initial state of a component, you can now display anything that's inside it in the UI.

You can render within the render method a value of any property that is in the state.

Define an h1 tag in the render method of the component the value of the name property, use the curly braces to access the state value followed by the property name, this will render the string Freecodecamp on the user's screen.