Render state in UI another way
There is another way to access state in a component. In the render() method, before the return statement, you can write Javascript directly.

For example, declaring functions or accessing prop data or also doing calculations on that data. You can then assign that expression to a variable.

In the render method, define a constant variable called name and set it equal to the value of the name property that is inside the component's state, you don't need to use the curly braces to assign this expression in the variable because you are writing Javascript code there.

The myComponent should render the value an h1 header element with the rendered text of the class component's state, enclosed in a div.