Set the component's state with this.setState
There is also another way to change the state of a component. React provides a method that allows us to update the state of the component, this method is called setState.

You call the setState method inside the component class as follows: this.setState(), inside the parentheses you pass it as an object having key-value pairs. The keys are the properties within the state and the values are the updated state data.

React expects you to never modify the state directly, instead using this.setState() when state changes occur.

There is a button element in our code that has an event called onClick( ). This event is called when the button receives a click event and executes the handleClick method that was defined in the component.

Inside the handleClick() method, update the component's state using this.setState(). Set the name property on state to equal the string React Rocks!.

Then, when the button is clicked by the user, it will update the h1 header element that had the Initial State text to React Rocks.