Bind the 'this' to a class method
In addition to defining and updating states, you can also define methods for the component class. A class method usually needs to use the keyword 'this' in front of the method name in order to be able to access the properties of the class within the scope of that method.

There are a few ways to allow class methods to access this, one way to do this is to explicitly bind this within the constructor function so that it is bound to the class methods when the component is initialized.

Then add a click event called onClick to the button element in the render method, this onClick event, should add the handleClick() method when the button receives a click event.

The method we pass to the onClick event needs to be enclosed in curly braces because it must be interpreted directly as Javascript.

Then when the user clicks the button the header will update and show the text You clicked! on the screen.