Use state to toggle an element's state
Sometimes you may need to know the previous state to update the state. However, state updates can be asynchronous, which means that React can bundle multiple setState( ) calls into a single update.

You can use a function inside setState that allows you to access the state or prop and ensure that you are working with the most current state or props values.

MyComponente has a property called visibility initialized with the boolean value false inside the state.

Make the value of this visibility property toggle its state between true and false. Pass a function to setState to define this method so that the visibilty state toggles to the opposite value when the method is called. If visibility is false, the method will set it to true and vice versa.

There is a click handler on the button elements by adding the toggleVisibility() method.

If the visibility is false, the visibility will be converted to true and the header with the text Now you see me! , if true it will convert to false and hide the header.