Create a simple counter in React
Let's create a more complex stateful component, a component that starts with a state, methods that set the state, and assign click events to add those methods.

The Counter component controls the state count value, create the increment( ) and decrement( ) methods so that the counter value is incremented or decremented by 1 when the appropriate button is clicked, also create the reset() method so that when the reset button is clicked, the count goes back to the beginning of the number 0.

So when the Increment button receives a click event, it will call the increment method and the counter will be incremented by 1.

When the Decrement button receives a click event, it will call the decrement method and decrement the counter by 1.

When the Reset button receives the click event, it will call the reset method and the counter value will return to 0.