Create a controlled input
Your application may have multiple interactions on an input field, for example, the user may be typing their data in the input and you want the data that the user typed to also be part of the state in the component. With React, you can make the value of the data entered in the field by the user the same as the application component, so with React you have the power to control the value of the input field.

Within the ControlledInput component, specifically in the state, there is an input property that starts with the value of an empty string, this value represents the text that the user enters in the field. It starts empty, we want that when the user starts writing in the field that the value of the input changes its state.

Create a method called handleChange() that has a parameter called event, when the method is called it will receive an event object that will contain the text string of the input element. We'll access this string with the event.target.value inside the method and update the state's input property value with this new string.

In the render method, create an input element above the h4 tag. In this input element add a value attribute that is equal to the component's input property. Then add the event handler named onChange( ) defined to the handleChange() method.

When the user starts writing in the input box, that text will be processed by the handleChange() method, set to the value of the input property in the component state, and rendered as the value in the input box on the page.