Create a controlled form
The form element is one of the elements that we can control the internal state.

The MyForm component is configured with an empty form with a submit handler. The submission handler will be called when the form is submitted.

Add an input element with the value attribute with its value being equal to the text that the user enters and also add the onChange attribute.

You must then complete the handleSubmit method so that it sets the state property of the submit to the input's current value.

Inside the handleSubmit method, create a method called event.preventDefault( ) to prevent the default form submission behavior.

Finally, create an h1 tag after the form's closing tag that renders the component's state submission value.

When the user writes something in the input field and clicks the button, he will see what he wrote in the input field on the page.