Pass state as props to child component
React allows us that the child components have the same states as a parent component and then rendering the child components, if you want the child components to have access to some parts of that state, you pass props in the component.

We have a component called MyApp in our code that renders Navbar child component. Pass the name property in state to the child component and show the name in the h1 tag that is part of the Navbar render method.

When the page renders, it will show the header text Hello, my name is: CamperBot.