Pass a function as props to other components
React allows you to pass event handlers or methods defined in a component to a child component, this is how you allow child components to interact with their parent components, you assign a name and access that method with this.props, that props will be the name of the function you want to pass.

In the code, we have a MyApp parent component that will render the GetInput and RenderInput child components. Add the GetInput component to the MyApp render method, then pass it a prop called input, assigning the MyApp state inputValue. Still inside this component, also create a prop called handleChange and pass the handleChange input handler to it as props.

Add the RenderInput component to the render method in MyApp, pass it a prop called input, and give the input the state's inputValue.

When you're done doing this whole process, you can type in the input field in the GetInput component that will call the handler method of the parent component through props. This will update the input value in state.

The user will be there typing in the GetInput field and each time he types what he wrote will appear rendered in the InputRender.