Use componentWillMount() lifecycle method on component
React components have several special methods that provide opportunities to perform actions at specific points in a component's lifecycle, these methods are called lifecycle methods.

If you want to perform some action before the component is rendered in the page's DOM, use the componentWillMount() method. It allows you to perform an action before the component is fully mounted.

After the method is executed the component will create the node in the browser. The componentWillMount() method is called before the render() method.

Register something in the console inside the componentWillMount() method.