Use the componentDidMount lifecycle method on a component
Most web developers will at some point need to call an API endpoint to retrieve data. When you are using React it is always important to know where you are going to perform this action.

The best practice in React is to make API calls or any other calls to the server in the componentDidMount() lifecycle method, this method will be executed after the component is mounted in the DOM.

Any setState( ) call here will trigger a re-rendering of the component, when you call an API within the method and set its state to the data the API returns, it will automatically trigger an update of the data as soon as you receive the data.

There is an API call inside the componentDidMount( ) method, it is setting the state after 2.5 seconds to simulate calling a server to retrieve data. This example requests the current total active users for a site.

Inside the render method, render the current value of activeUsers in the h1 element along with the text Active Users:. So after the component is rendered the current value of users will be shown on the page after 2.5 seconds.