Add event listeners inside componentDidMount() method and remove in componentWillUnmount() method
The componentDidMount() method is the best place to add any event listeners you want to use for specific functionality. React provides a synthetic event system that wraps around the native event system already present in browsers.

Some synthetic event handlers we are already using, like onClick() for example. However, if you want to attach an event handler to document objects, you must do so directly in the document or window.

Attach an event listener to the componentDidMount() method, use document.addEventListener() passing the keydown event between quotes and the second argument will be the function that will be triggered when listening to this event.

Then, in componentWillUnmount() remove that event listener, passing the same arguments to document.removeEventListener(). It's good practice to use this component lifecycle methodWillUnmount( ) to clean up React components before they are unmounted and destroyed.

So when the user presses the enter key on his keyboard, he will see the header text 'You pressed the enter key!' on the page screen.