Optimize Re-Renders with the shouldComponentUpdate lifecycle method
React provides a lifecycle method called shouldComponentUpdate(), it is only called when the child components receive a new state or props and it states a question 'Was the component updated?', if so the component is updated and re-rendered , otherwise it is ignored and remains in the same state. The shouldComponentUpdate() method takes nextProps and nextState as parameters.

This method is a way for you to optimize the performance of the component. The default behavior is that whenever it receives new props and has new changes, the component is rendered again, with shouldComponentUpdate() you can avoid this by comparing the props.

Inside the shouldComponentUpdate() method, compare the current props (this.props) with the next props (nextProps) to determine if you need to update or not.

The shouldComponentUpdate() method is added to a component called OnlyEvens. Currently, this method returns true, causing the component to re-render every time it receives new props.

Modify the shouldComponentUpdate() method so that OnlyEvens is rendered only if the value of your new props are even.

So when the value of nextProps is an even number, OnlyEvens will be rendered.