Give inline styles to JSX element
Another powerful feature of React is that we can style our JSX elements.

When you style JSX elements in React, you need to use a style attribute like in HTML, only due to the way JSX is transpiled, we cannot set the style value to a string, instead you set it equal to an object in Javascript using the braces and within the braces the styles for that element.

Add a style attribute to the div element in the code editor to give the text a red color and a font size of 72px. You don't need to set the font size with the px unit, you can only set it with the number 72.