Add inline styles to a variable
See the syntax difference of inline HTML styles for JSX elements, plus the style attribute value is set to a Javascript object. Other differences are noted here as the names of certain CSS style properties are case sensitive.

For example, the property that defines text size in React is fontSize and not font-size as used in CSS because hyphenated words like font-size are invalid syntax for Javascript object properties, so React uses camel case.

Another difference is the length units of property values such as (height, width and fontSize) are considered in px, there is also another way of giving value to these properties which is using 'em', all values of this property need to be between the quotes.

If you have a large set of styles, you can assign a styles object to a constant variable. To keep your code organized, declare a style constant variable as a global variable at the top of your code. Inside the styles constant, define the color, fontSize and border properties.

In the color property, set the value to purple, in the fontSize property, set the value to 40 and in the border property, set the value 2px solid purple.

Give the div a purple color, a font size of 40, and a 2px border of solid purple.