Use advanced Javascript in React component's render method
You can put Javascript directly inside the render method, before the return statement, without inserting them between the curly braces. This is allowed because it is not inside the JSX element, but if you want to use Javascript inside the return statement in the JSX element, you must put the variable name between the curly braces.

In our code, we have the render method, the first thing inside it is an array containing 20 sentences to represent the answers found in the classic Magic Eight Ball toy from the 80s. We have the button click event linked to the ask method, therefore, each time the for button is clicked a random number will be generated and stored as the randomIndex in the state.

Remove on line 52 the string change me! of the answer constant variable and reassign the answer so that the code randomly accesses a different index of the possibleAnswers array whenever the component is updated. Finally, inside the p tags insert the answer constant.

So every time the user writes some text question in the input field and clicks on the button, he will see a random answer from the magic eight ball rendered on the screen.