Render the component with If-Else conditions
Another use of Javascript is to bind the rendered elements that are rendered to a condition, this allows you to control which elements will be rendered and displayed on the screen. When the condition is true, a view is rendered, when false, another view is rendered.

We can do this with an if/else statement inside the render( ) method of the React component. The Mycomponent contains within its state a boolean to track whether the user wants to display some element on the screen. The button toggles the state of that value.

Currently, the state value is true, making the header element with the text Displayed! is always rendered and shown on the screen.

Rewrite the render( ) method with an if/else statement so that if the display is true, a div, button, and the h1 are rendered, if the display is false, only a div and the button are rendered.

So, for example, the initial display state is false, when the user clicks on the button the display will become true and will show the text Displayed on the user's screen! as header.