Use the && operator to make a more concise conditional
There's another way we can make conditions to get the same result, just more concisely. There is a condition that is also used in Javascript that allows you to render an element if the condition is true, if the condition is false the element is not rendered.

Instead of using multiple if-else statements, you can use the logical && operator to perform more concise conditional logic. For example, we want if a condition is true, a tag is returned. We would use it like this:

{condition && <p>markup</p>}

If the condition is true, the tag is returned. If the condition is false, nothing is returned. This allows us to handle more complex conditional logic within the render() method without repeating too much code.

Use the && operator so that the h1 is rendered only if the display is true, if it is false then only a div and the button should be rendered.